refactor: relocate organism contract to a site-neutral home (#1411)
The structural `Organism` contract consumed by the shared OrganismView and
workflow side panel lived in app/views/OrganismView/types.ts — a view
directory imported by shared viewModelBuilders and other views. Move it to a
site-neutral home, app/apis/catalog/common/entities.ts, renamed
OrganismContract to avoid the name clash with the (still-used) union in
WorkflowsView/types.ts. OrganismView/types.ts keeps a transitional shim
re-export so existing `Organism` importers are unchanged.

Part of #1246.
